+++ mlrun/launcher/client/launcher.py
@@ -11,10 +11,10 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-import mlrun.common.runtimes
+import mlrun.common.launcher
 
 
-class ClientLocalLauncher(mlrun.common.runtimes.AbstractLauncher):
+class ClientLocalLauncher(mlrun.common.launcher.AbstractLauncher):
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
 
@@ -35,4 +35,31 @@ class ClientLocalLauncher(mlrun.common.runtimes.AbstractLauncher):
         pass
 
     def _run_local(self, runtime):
+        pass
+
+
+class ClientRemoteLauncher(mlrun.common.runtimes.AbstractLauncher):
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+
+    @staticmethod
+    def verify_base_image(runtime):
+        pass
+
+    @staticmethod
+    def save(runtime):
+        pass
+
+    @staticmethod
+    def run(runtime):
+        pass
+
+    @staticmethod
+    def _enrich_and_validate(runtime):
+        pass
+
+    def _submit_job(self, runtime):
+        pass
+
+    def _deploy(self, runtime):
         pass
